feat!: re-export functionality as a library
